+++ src/org/robockets/robot/cubeintake/CubeIntake.java
@@ -1,4 +1,4 @@
-package org.robockets.robot.gearintake;
+package org.robockets.robot.cubeintake;
 
 import org.robockets.commons.RelativeDirection;
 import org.robockets.robot.RobotMap;
@@ -8,7 +8,7 @@ import edu.wpi.first.wpilibj.command.Subsystem;
 /**
  *
  */
-public class  GearIntake extends Subsystem {
+public class CubeIntake extends Subsystem {
 
     // Put methods for controlling this subsystem
     // here. Call these from Commands.
@@ -18,8 +18,8 @@ public class  GearIntake extends Subsystem {
         //setDefaultCommand(new MySpecialCommand());
     }
     
-    public void moveIntakeArm(RelativeDirection.XAxis direction,double speed){
-    	if (direction == RelativeDirection.XAxis.LEFT){
+    public void moveIntakeArm(RelativeDirection.XAxis direction, double speed){
+			if (direction == RelativeDirection.XAxis.LEFT){
     		RobotMap.leftIntake.set(speed);
     	} else {
     		RobotMap.rightIntake.set(speed);
